Update EntityIdPool.cpp
Added some tests to make it easier to check if an entity exists at a given index without throwing asserts.
Removed a resize command in the 'clear' function. The code seems to expect the m_counts vector and several other vectors to be the same size. When the World::clear function is called, this call to 'resize' breaks that assumption, which can result in crashes due to out-of-bounds indexes on the other vectors.